Refactor: move http and url utilities to net/
Moved http-client, http-status, and url modules from lib/ to net/ and updated all imports accordingly. This improves code organization by grouping network-related utilities under the net/ directory.
