+++ src/errors/registry.ts
@@ -1,4 +1,4 @@
-import HttpStatus from '@/lib/http-status'
+import HttpStatus from '@/net/http/status'
 
 import ErrorCode from './codes'
