+++ src/routes/auth/logout/__tests__/endpoint.test.ts
@@ -3,7 +3,7 @@ import type { Hono } from 'hono'
 import { afterEach, beforeEach, describe, expect, it, mock } from 'bun:test'
 
 import { createTestApp, doRequest, ModuleMocker, post } from '@/__tests__'
-import HttpStatus from '@/lib/http-status'
+import HttpStatus from '@/net/http/status'
 
 import logoutRoute from '../index'
