+++ src/routes/auth/request-password-reset/handler.ts
@@ -1,6 +1,6 @@
 import type { Context } from 'hono'
 
-import HttpStatus from '@/lib/http-status'
+import HttpStatus from '@/net/http/status'
 
 import type { RequestPasswordResetBody } from './schema'
