+++ src/routes/auth/resend-verification-email/__tests__/endpoint.test.ts
@@ -3,8 +3,8 @@ import type { Hono } from 'hono'
 import { afterEach, beforeEach, describe, expect, it, mock } from 'bun:test'
 
 import { createTestApp, doRequest, ModuleMocker, post } from '@/__tests__'
-import HttpStatus from '@/lib/http-status'
 import { mockCaptchaSuccess, VALID_CAPTCHA_TOKEN } from '@/middleware/captcha/__tests__'
+import HttpStatus from '@/net/http/status'
 
 import resendVerificationEmailRoute from '../index'
