+++ src/routes/auth/reset-password/handler.ts
@@ -1,6 +1,6 @@
 import type { Context } from 'hono'
 
-import HttpStatus from '@/lib/http-status'
+import HttpStatus from '@/net/http/status'
 
 import type { ResetPasswordBody } from './schema'
