+++ src/routes/auth/verify-email/handler.ts
@@ -1,7 +1,7 @@
 import type { Context } from 'hono'
 
 import { setAccessCookie } from '@/lib/cookie'
-import HttpStatus from '@/lib/http-status'
+import HttpStatus from '@/net/http/status'
 
 import type { VerifyEmailBody } from './schema'
